Add CASE(...) in favour of TEST(...)
According to Kevlin Henney in "Making Steaks from Sacred Cows"[1], a test case should correspond to a single case, it should have have a single reason to change. For this reason it's good to use the phrase test case. As it'sclear we're talking about tests, I omit the word "test".

[1] https://vimeo.com/105758303, from ca. 0:58